# swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "Cinemax API",
    "description": "API документация Cinemax приложения"
  },
...
  "paths": {
...
    "/api/movies": {
      "get": {
        "summary": "Получить список всех фильмов",
        "operationId": "getMovies",
        "tags": ["Movies"],
        "responses": {
          "200": {
            "description": "Список фильмов успешно получен"
          }
        }
      },
      "post": {
        "summary": "Создать новый фильм",
        "operationId": "createMovie",
        "tags": ["Movies"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Фильм успешно создан"
          }
        }
      }
    },
    "/api/movies/{id}": {
      "get": {
        "summary": "Получить фильм по ID",
        "operationId": "getMovie",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные фильма успешно получены"
          }
        }
      },
      "patch": {
        "summary": "Обновить фильм",
        "operationId": "updateMovie",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Фильм успешно обновлен"
          }
        }
      },
      "delete": {
        "summary": "Удалить фильм",
        "operationId": "deleteMovie",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Фильм успешно удален"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "main_img": {
            "type": "string",
            "format": "binary",
            "description": "Файл изображения"
          },
          "horizontal_img": {
            "type": "string",
            "format": "binary",
            "description": "Файл изображения"
          },
          "video": {
            "type": "string",
            "format": "binary",
            "description": "Файл видео"
          },
          "name": {
            "type": "string"
          },
          "descr": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "nullable": true
          },
          "year_of_creation": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "genre": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          }
        }
      }
    }
  }
}